+++ paper/paper.bib
@@ -6,7 +6,7 @@
                and Markov, Nikolay S and Zappia, Luke and Luecken, Malte D and
                Interlandi, Marta and Saeys, Yvan and Saelens, Wouter",
   journal   = "Journal of Open Source Software",
-  doi       = "10.21105.joss.07698",
+  doi       = "10.21105/joss.07698",
   volume    =  10,
   number    =  108,
   pages     =  7698,
