### 2026-02-20 e96ffa2
Add OMB reference to bib file
