+++ paper/paper.bib
@@ -78,3 +78,15 @@
   url       = "https://CRAN.R-project.org/package=shiny",
   doi       = "10.32614/CRAN.package.shiny",
 }
+
+@ARTICLE{mallona2026-omnibenchmark,
+  title     = "Omnibenchmark: transparent, reproducible, extensible and
+               standardized orchestration of solo and collaborative
+               benchmarks",
+  author    = "Mallona, Izaskun and Luetge, Almut and Carrillo, Ben and
+               Incicau, Daniel and Gerber, Reto and Meara, Aidan and Sonrel,
+               Anthony and Soneson, Charlotte and Robinson, Mark D",
+  journal   = "arXiv [q-bio.OT]",
+  year      =  2026,
+  doi       = "10.48550/arXiv.2409.17038"
+}
