### 2026-02-20 2d85d8f
Expand paper adapting to new JOSS guidelines
