+++ paper/paper.md
@@ -83,18 +83,46 @@ For example, a user working locally in R may find it more convenient to generate
 _bettr_ can also be deployed in server mode, allowing users to upload their own input data (in JSON format) to a running app. 
 
 
-# Statement of need
+# Statement of Need
 
 The interactive nature of _bettr_ makes it particularly suitable for visual, exploratory analysis of benchmarking results, since the user can interactively modify the weights associated with the different evaluation metrics and immediately see how it affects the ranking of the methods. 
 It is also straightforward for authors of benchmarking studies to deploy an instance of _bettr_ using, e.g., a local Shiny server or a commercial option such as [https://shinyapps.io](https://shinyapps.io), to allow readers to easily explore their results. 
-Complementary functionality for creating static summary representations (e.g., heatmap-like visualizations) of benchmarking results is provided e.g. by the funkyheatmap package [@cannoodt2025-funkyheatmap], and benchmarking platforms such as OpenEBench [@capella2017-openebench] and OpenProblems [@Luecken2025-openproblems] also produce result visualizations for the included benchmarks. 
-However, existing tools typically lack either flexibility in input format or means of deployment, or the ability to explore results interactively, and are not intended to support user-specific metric weighting. 
 
-# Availability and examples
+# State of the Field
+
+Complementary functionality for creating summary representations (e.g., heatmap-like visualizations) of benchmarking results is provided e.g. by the funkyheatmap package [@cannoodt2025-funkyheatmap], and benchmarking platforms such as OpenEBench [@capella2017-openebench] and OpenProblems [@Luecken2025-openproblems] also produce result visualizations for the included benchmarks. 
+However, existing tools typically lack flexibility in either input format or means of deployment, or the ability to explore results interactively, and are not intended to support user-specific metric weighting.
+To the best of our knowledge, _bettr_ is the first generic tool for benchmark visualization that combines interactivity and flexibility in metric weighting with ease of use and a transparent, programmatic interface. 
+
+# Software Design
+
+The design philosophy behind _bettr_ focuses on flexibility, accessibility, and reproducibility. 
+By supporting several input formats, from collections of data frames via a single R-based object to an all-encompassing JSON file, _bettr_ can be used for downstream analysis and visualization of results generated using a wide range of benchmark setups and programming languages. 
+_bettr_ is fully open source and easily installable as an R package, mainly distributed via Bioconductor with the most recent development version also available on GitHub.
+In addition, it is available via [_r-universe_](https://bioc.r-universe.dev/bettr), which among other things provides binaries for several platforms, including WebAssembly.
+As a complement to the interactive interface, full reproducibility is enabled by the equivalent programmatic interface to the functionality, as well as the ability to export the processed data as either a shareable csv file containing the metric values as well as the final score, or an R list that can be used directly as the input for further analysis and visualization.
+
+# Research Impact Statement
+
+In 2025, _bettr_ was [downloaded](https://bioconductor.org/packages/stats/bioc/bettr/) almost 3,000 times (by 1,593 unique IPs) from Bioconductor alone. 
+Adoption of _bettr_ is expected to increase further in the near future, as it has become the first supported integration for automated metrics reporting within the [Omnibenchmark](https://www.omnibenchmark.org) project. 
+Omnibenchmark [@mallona2026-omnibenchmark] is a benchmarking system that automates and standardizes routine aspects of benchmarking through standardisation and formalisation of benchmarking plans. 
+During execution, Omnibenchmark collects computational performance metrics (e.g., peak memory usage, CPU utilization, runtime, etc), as well as, where applicable, algorithmic performance metrics (e.g., F1 scores, ARIs, etc). 
[…]
