+++ articles/bettr.md
@@ -771,13 +771,13 @@ encourage users to contribute to that via a Pull Request to
 
 ``` r
 sessionInfo()
-#> R Under development (unstable) (2025-11-24 r89053)
-#> Platform: aarch64-apple-darwin20
-#> Running under: macOS Sequoia 15.7.2
+#> R Under development (unstable) (2026-03-20 r89666)
+#> Platform: aarch64-apple-darwin23
+#> Running under: macOS Sequoia 15.7.4
 #> 
 #> Matrix products: default
-#> BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
-#> LAPACK: /Library/Frameworks/R.framework/Versions/4.6-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.1
+#> BLAS:   /Library/Frameworks/R.framework/Versions/4.6/Resources/lib/libRblas.0.dylib 
+#> LAPACK: /Library/Frameworks/R.framework/Versions/4.6/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.1
 #> 
 #> locale:
 #> [1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8
@@ -790,46 +790,46 @@ sessionInfo()
 #> [8] base     
 #> 
 #> other attached packages:
-#>  [1] dplyr_1.1.4                 tibble_3.3.0               
-#>  [3] SummarizedExperiment_1.41.0 Biobase_2.71.0             
-#>  [5] GenomicRanges_1.63.0        Seqinfo_1.1.0              
+#>  [1] dplyr_1.2.0                 tibble_3.3.1               
+#>  [3] SummarizedExperiment_1.41.1 Biobase_2.71.0             
+#>  [5] GenomicRanges_1.63.1        Seqinfo_1.1.0              
 #>  [7] IRanges_2.45.0              S4Vectors_0.49.0           
 #>  [9] BiocGenerics_0.57.0         generics_0.1.4             
 #> [11] MatrixGenerics_1.23.0       matrixStats_1.5.0          
 #> [13] bettr_1.7.1                 BiocStyle_2.39.0           
 #> 
 #> loaded via a namespace (and not attached):
-#>  [1] gridExtra_2.3         rlang_1.1.6           magrittr_2.0.4       
-#>  [4] clue_0.3-66           GetoptLong_1.1.0      otel_0.2.0           
-#>  [7] compiler_4.6.0        png_0.1-8             systemfonts_1.3.1    
-#> [10] vctrs_0.6.5           stringr_1.6.0         pkgconfig_2.0.3      
+#>  [1] gridExtra_2.3         rlang_1.1.7           magrittr_2.0.4       
+#>  [4] clue_0.3-67           GetoptLong_1.1.0      otel_0.2.0           
+#>  [7] compiler_4.6.0        png_0.1-9             systemfonts_1.3.2    
+#> [10] vctrs_0.7.1           stringr_1.6.0         pkgconfig_2.0.3      
 #> [13] shape_1.4.6.1         crayon_1.5.3          fastmap_1.2.0        
 #> [16] backports_1.5.0       XVector_0.51.0        labeling_0.4.3       
 #> [19] utf8_1.2.6            learnr_0.11.6         shinyjqui_0.4.1      
-#> [22] promises_1.5.0        rmarkdown_2.30        ragg_1.5.0           
-#> [25] purrr_1.2.0           xfun_0.54             cachem_1.1.0         
-#> [28] jsonlite_2.0.0        later_1.4.4           DelayedArray_0.37.0  
[…]
