+++ articles/server-mode.md
@@ -437,13 +437,13 @@ When version changes:
 
 ``` r
 sessionInfo()
-#> R Under development (unstable) (2025-11-24 r89053)
-#> Platform: aarch64-apple-darwin20
-#> Running under: macOS Sequoia 15.7.2
+#> R Under development (unstable) (2026-03-20 r89666)
+#> Platform: aarch64-apple-darwin23
+#> Running under: macOS Sequoia 15.7.4
 #> 
 #> Matrix products: default
-#> BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
-#> LAPACK: /Library/Frameworks/R.framework/Versions/4.6-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.1
+#> BLAS:   /Library/Frameworks/R.framework/Versions/4.6/Resources/lib/libRblas.0.dylib 
+#> LAPACK: /Library/Frameworks/R.framework/Versions/4.6/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.1
 #> 
 #> locale:
 #> [1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8
@@ -459,13 +459,13 @@ sessionInfo()
 #> 
 #> loaded via a namespace (and not attached):
 #>  [1] digest_0.6.39       desc_1.4.3          R6_2.6.1           
-#>  [4] bookdown_0.45       fastmap_1.2.0       xfun_0.54          
-#>  [7] cachem_1.1.0        knitr_1.50          htmltools_0.5.8.1  
-#> [10] rmarkdown_2.30      lifecycle_1.0.4     cli_3.6.5          
-#> [13] sass_0.4.10         pkgdown_2.2.0.9000  textshaping_1.0.4  
-#> [16] jquerylib_0.1.4     systemfonts_1.3.1   compiler_4.6.0     
-#> [19] tools_4.6.0         ragg_1.5.0          bslib_0.9.0        
-#> [22] evaluate_1.0.5      yaml_2.3.11         BiocManager_1.30.26
-#> [25] jsonlite_2.0.0      rlang_1.1.6         fs_1.6.6           
-#> [28] htmlwidgets_1.6.4
+#>  [4] bookdown_0.46       fastmap_1.2.0       xfun_0.57          
+#>  [7] cachem_1.1.0        knitr_1.51          htmltools_0.5.9    
+#> [10] rmarkdown_2.30      lifecycle_1.0.5     cli_3.6.5          
+#> [13] sass_0.4.10         pkgdown_2.2.0.9000  textshaping_1.0.5  
+#> [16] jquerylib_0.1.4     systemfonts_1.3.2   compiler_4.6.0     
+#> [19] tools_4.6.0         ragg_1.5.1          bslib_0.10.0       
+#> [22] evaluate_1.0.5      yaml_2.3.12         BiocManager_1.30.27
+#> [25] otel_0.2.0          jsonlite_2.0.0      rlang_1.1.7        
+#> [28] fs_1.6.7            htmlwidgets_1.6.4
 ```
