+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 3.1.11
 pkgdown: 2.2.0.9000
-pkgdown_sha: c07d935583c90125586821e3b39ed52f16593e42
+pkgdown_sha: a6abe43ed48e58b1de42eba5e5541a5b6c46b46f
 articles:
   bettr: bettr.html
   server-mode: server-mode.html
-last_built: 2025-11-29T17:31Z
+last_built: 2026-03-21T09:56Z
 urls:
   reference: https://federicomarini.github.io/bettr/reference
   article: https://federicomarini.github.io/bettr/articles
