### 2026-03-31 #9661 @crvernon
👋 Alright @csoneson, @srvanderplas, @robertmartin8, @rcannood - I am going to close this Pre-Review and kick off the full review which you should receive a notification for. Thanks!
