### 2026-03-31 #10333 @editorialbot
[REVIEW]: Interactive exploration of benchmarking results with bettr
**Submitting author:** @csoneson (<a href="http://orcid.org/0000-0003-3833-2169">Charlotte Soneson</a>)
**Repository:** https://github.com/federicomarini/bettr/
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.7.1
