### 2026-03-31 #10333 @crvernon
👋 @csoneson, @srvanderplas, @robertmartin8, @rcannood  - This is the review thread for the paper. All of our communications will happen here from now on.
