### 2026-03-31 #10333 @srvanderplas
- Not actually stated in the license, but it's the text of the MIT license.
