### 2026-04-01 federicomarini/bettr#27 @robertmartin8
"Highlight ID" is only implemented for Parallel coordinates
`highlightMethod` is only wired to the parallel coordinates plot, but the UI places the "Highlight ID" control in the sidebar where it appears on every tab, giving the impression it should affect all views. On the heatmap, polar plot, and bar tabs it does nothing.

Here's a minimal reproducer

```R
  library(bettr)

  df <- data.frame(
    Method = c("A", "B", "C"),
    metric1 = c(0.9, 0.5, 0.7),
    metric2 = c(0.4, 0.8, 0.6)
  )

  bettr(df = df, idCol = "Method")
  # Steps to reproduce:
  # 1. In the sidebar, select any method from "Highlight ID"
  # 2. Observe: no visual change on Heatmap, Polar, or Bar plot tabs
  # 3. Switch to "Parallel coordinates" tab — highlight works correctly there
```

**Expected**: Either the highlight applies across all plot types, or the "Highlight ID" control is only shown when the Parallel coordinates tab is active.

 **Actual**: The control is visible on all tabs but only has an effect on Parallel coordinates.

https://github.com/openjournals/joss-reviews/issues/10333
