### 2026-04-01 #10333 @robertmartin8
- I am willing to say yes on the basis of the [Omnibenchmark](https://docs.omnibenchmark.org/latest/news/) integration.

    - while there are no external developers, there appear to be 26 PRs across 3 contributors.
    - the project does have good documentation and examples though these seem a bit scattered. I have raised an issue accordingly.

    -  `BiocManager::install` worked for me
    - was able to get the shiny app working

     - [x] as noted in the issue I raised, it would be best if these were explicitly linked
    - good test coverage and I successfully ran them all locally.

    - I looked at some of the other mentioned tools and concluded that `bettr` does have features that are not present therein.
    - I judge that there is "credible near-term significance"  because of the 1593 unique downloads, relative to my guess of the size of this subfield field (bioinformatics benchmarking).
