### 2026-04-01 federicomarini/bettr#28 @robertmartin8
Suggestion: better link to documentation and examples
instead of just hyperlinking to "vignettes", I think it would be better if you made a documentation button, or at least moved that to the top of the readme.

Also, I note you have the "bettr-examples" repository. This is quite helpful and should be linked in the readme.

https://github.com/openjournals/joss-reviews/issues/10333
