### 2026-04-01 #10333 @robertmartin8
Hi @csoneson.  have conducted my initial review of the submission and the associated software. I believe there there are several points that need to be addressed before the manuscript meets the JOSS publication standards.

1. **Authorship and contributions**: the submission lists nine co-authors, but only three appear to have made substantive contributions to the codebase based on the commit history. As per JOSS authorship guidelines, code contributions are _not_ a necessary condition for authorship, but it may be helpful to confirm here that the co-authors do meet the below criteria.

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but **active project direction and other forms of non-code contributions are**

2. **Summary**: the summary is very long and it intermixes high-level motivation with implementation-level detail (like specific algorithm choices and architectural decisions). I recommend the authors revise the summary to provide a concise, accessible overview to a nonspecialist audience (as per JOSS guidelines). Some of the verbiage here may be better placed in the Statement of Need section (see below).

3. **Statement of Need**: in my judgment, the Statement of Need does not adequately address the JOSS guidelines, in particular the "research purpose of the software" and the "target audience".

In addition to the above, I have raised two specific code suggestions in the software repository. While these are not blockers, I think they would improve the usability of the software.

Best,
Robert
