### 2026-04-04 37d273e
Add quick links section to README
