### 2026-04-04 1aaf37b
Move highlightMethodUI to the parallel coordinates tab, allow multiple selected values
