### 2026-04-04 federicomarini/bettr#28 @csoneson
Thanks for the suggestion - I have added a quick links section to the README in 37d273e85c9aed48d34ec52963be83756e77696b
