### 2026-04-04 federicomarini/bettr#27 @csoneson
Good point - I moved the highlight method UI to the parallel coordinates tab in 1aaf37bfa74542bed16e71b39b2c5d7883622b39
