### 2026-04-04 #10333 @csoneson
@robertmartin8 Thanks for your comments!

Regarding (1), I confirm that while not all authors have explicitly written the code that went into `bettr`, all have contributed significantly to the development of the package over time, e.g. via extensive discussions of the scope, content and interface, testing, feature requests, deployment and integration with other tools.

We will report back here when we have addressed the other manuscript-related comments (the two issues in the software repo have been fixed).
