+++ paper/paper.bib
@@ -90,3 +90,20 @@
   year      =  2026,
   doi       = "10.48550/arXiv.2409.17038"
 }
+
+@ARTICLE{Sonrel2023-metaanalysis,
+  title     = "Meta-analysis of (single-cell method) benchmarks reveals the need
+               for extensibility and interoperability",
+  author    = "Sonrel, Anthony and Luetge, Almut and Soneson, Charlotte and
+               Mallona, Izaskun and Germain, Pierre-Luc and Knyazev, Sergey and
+               Gilis, Jeroen and Gerber, Reto and Seurinck, Ruth and Paul,
+               Dominique and Sonder, Emanuel and Crowell, Helena L and
+               Fanaswala, Imran and Al-Ajami, Ahmad and Heidari, Elyas and
+               Schmeing, Stephan and Milosavljevic, Stefan and Saeys, Yvan and
+               Mangul, Serghei and Robinson, Mark D",
+  journal   = "Genome Biology",
+  volume    =  24,
+  number    =  1,
+  pages     =  119,
+  year      =  2023
+}
