### 2026-04-16 #10333 @robertmartin8
HI @csoneson, I believe there is still substantial duplication between the Summary and Statement of Need.

Both sections describe the general role of benchmarking, the fact that users may value metrics differently, and the interactive reweighting/exploration enabled by bettr. I would suggest separating these sections more clearly.

The Summary should stay focused on a concise, high-level description of what bettr is, what it does, and who it is for, for a broad JOSS audience. The Statement of Need should then focus more specifically on the gap bettr fills: for example, the difficulty of exploring benchmarking results beyond a particular configuration, the lack of readily explorable benchmark outputs, and why existing workflows or tools are insufficient.

As written, parts of the Statement of Need read like a second summary of the package rather than a distinct justification for why the software is needed.

Hope I'm not coming across as too pedantic, I happen to think those two sections are quite important.

Best,
Robert
