+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: bettr
 Title: A Better Way To Explore What Is Best
-Version: 1.8.0
+Version: 1.9.0
 Date: 2025-11-16
 Authors@R: 
     c(person("Federico", "Marini", 
