+++ articles/bettr.md
@@ -771,7 +771,7 @@ encourage users to contribute to that via a Pull Request to
 
 ``` r
 sessionInfo()
-#> R version 4.6.0 alpha (2026-04-03 r89785)
+#> R version 4.6.0 Patched (2026-04-27 r89967)
 #> Platform: aarch64-apple-darwin23
 #> Running under: macOS Sequoia 15.7.4
 #> 
@@ -792,44 +792,44 @@ sessionInfo()
 #> other attached packages:
 #>  [1] dplyr_1.2.1                 tibble_3.3.1               
 #>  [3] SummarizedExperiment_1.41.1 Biobase_2.71.0             
-#>  [5] GenomicRanges_1.63.1        Seqinfo_1.1.0              
-#>  [7] IRanges_2.45.0              S4Vectors_0.49.0           
-#>  [9] BiocGenerics_0.57.0         generics_0.1.4             
+#>  [5] GenomicRanges_1.63.2        Seqinfo_1.1.0              
+#>  [7] IRanges_2.45.0              S4Vectors_0.49.2           
+#>  [9] BiocGenerics_0.57.1         generics_0.1.4             
 #> [11] MatrixGenerics_1.23.0       matrixStats_1.5.0          
-#> [13] bettr_1.7.2                 BiocStyle_2.39.0           
+#> [13] bettr_1.9.0                 BiocStyle_2.39.0           
 #> 
 #> loaded via a namespace (and not attached):
-#>  [1] gridExtra_2.3         rlang_1.1.7           magrittr_2.0.4       
-#>  [4] clue_0.3-68           GetoptLong_1.1.0      otel_0.2.0           
+#>  [1] gridExtra_2.3         rlang_1.2.0           magrittr_2.0.5       
+#>  [4] clue_0.3-68           GetoptLong_1.1.1      otel_0.2.0           
 #>  [7] compiler_4.6.0        png_0.1-9             systemfonts_1.3.2    
-#> [10] vctrs_0.7.2           stringr_1.6.0         pkgconfig_2.0.3      
+#> [10] vctrs_0.7.3           stringr_1.6.0         pkgconfig_2.0.3      
 #> [13] shape_1.4.6.1         crayon_1.5.3          fastmap_1.2.0        
 #> [16] backports_1.5.1       XVector_0.51.0        labeling_0.4.3       
 #> [19] utf8_1.2.6            learnr_0.11.6         shinyjqui_0.4.1      
 #> [22] promises_1.5.0        rmarkdown_2.31        ragg_1.5.2           
-#> [25] purrr_1.2.1           xfun_0.57             cachem_1.1.0         
+#> [25] purrr_1.2.2           xfun_0.57             cachem_1.1.0         
 #> [28] jsonlite_2.0.0        later_1.4.8           DelayedArray_0.37.1  
 #> [31] parallel_4.6.0        cluster_2.1.8.2       R6_2.6.1             
 #> [34] bslib_0.10.0          stringi_1.8.7         RColorBrewer_1.1-3   
 #> [37] rpart_4.1.27          jquerylib_0.1.4       assertthat_0.2.1     
-#> [40] Rcpp_1.1.1            bookdown_0.46         iterators_1.0.14     
+#> [40] Rcpp_1.1.1-1.1        bookdown_0.46         iterators_1.0.14     
 #> [43] knitr_1.51            base64enc_0.1-6       httpuv_1.6.17        
 #> [46] Matrix_1.7-5          nnet_7.3-20           tidyselect_1.2.1     
 #> [49] rstudioapi_0.18.0     abind_1.4-8           yaml_2.3.12          
[…]
