+++ articles/server-mode.md
@@ -437,7 +437,7 @@ When version changes:
 
 ``` r
 sessionInfo()
-#> R version 4.6.0 alpha (2026-04-03 r89785)
+#> R version 4.6.0 Patched (2026-04-27 r89967)
 #> Platform: aarch64-apple-darwin23
 #> Running under: macOS Sequoia 15.7.4
 #> 
@@ -461,11 +461,11 @@ sessionInfo()
 #>  [1] digest_0.6.39       desc_1.4.3          R6_2.6.1           
 #>  [4] bookdown_0.46       fastmap_1.2.0       xfun_0.57          
 #>  [7] cachem_1.1.0        knitr_1.51          htmltools_0.5.9    
-#> [10] rmarkdown_2.31      lifecycle_1.0.5     cli_3.6.5          
+#> [10] rmarkdown_2.31      lifecycle_1.0.5     cli_3.6.6          
 #> [13] sass_0.4.10         pkgdown_2.2.0.9000  textshaping_1.0.5  
 #> [16] jquerylib_0.1.4     systemfonts_1.3.2   compiler_4.6.0     
 #> [19] tools_4.6.0         ragg_1.5.2          bslib_0.10.0       
 #> [22] evaluate_1.0.5      yaml_2.3.12         BiocManager_1.30.27
-#> [25] otel_0.2.0          jsonlite_2.0.0      rlang_1.1.7        
-#> [28] fs_2.0.1            htmlwidgets_1.6.4
+#> [25] otel_0.2.0          jsonlite_2.0.0      rlang_1.2.0        
+#> [28] fs_2.1.0            htmlwidgets_1.6.4
 ```
