+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: a6abe43ed48e58b1de42eba5e5541a5b6c46b46f
 articles:
   bettr: bettr.html
   server-mode: server-mode.html
-last_built: 2026-04-04T12:18Z
+last_built: 2026-04-28T17:38Z
 urls:
   reference: https://federicomarini.github.io/bettr/reference
   article: https://federicomarini.github.io/bettr/articles
