+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: bettr
 Title: A Better Way To Explore What Is Best
-Version: 1.9.0
-Date: 2026-04-04
+Version: 1.9.1
+Date: 2026-09-12
 Authors@R: 
     c(person("Federico", "Marini", 
              role = "aut", email = "marinif@uni-mainz.de", 
