+++ paper/paper.bib
@@ -105,5 +105,6 @@
   volume    =  24,
   number    =  1,
   pages     =  119,
-  year      =  2023
+  year      =  2023,
+  doi       = "10.1186/s13059-023-02962-5"
 }
