### 2026-09-16 bc13b7f
Update paper
