+++ paper/paper.md
@@ -92,7 +92,7 @@ This has the potential to increase the usefulness of published benchmarks by dem
 Most published visualizations of benchmark results are created with custom scripts using generic R or python graphics libraries. 
 A small number of dedicated viewers for benchmark results exist, and _bettr_ was designed to complement their functionality, rather than to replace them. 
 The funkyheatmap R package [@cannoodt2025-funkyheatmap] generates elaborate, publication-ready static heatmaps of diverse types of benchmark results. 
-It is also wrapped into a standalone executable as well as a [nextflow module] (https://funkyheatmap.github.io/funkyheatmap/articles/nextflow.html), and the funkyheatmap.js [extension] (https://funkyheatmap.github.io/funkyheatmapjs/index.html) can be used to create visualizations that can be interactively browsed and sorted by a selected metric. 
+It is also wrapped into a standalone executable as well as a [nextflow module](https://funkyheatmap.github.io/funkyheatmap/articles/nextflow.html), and the funkyheatmap.js [extension](https://funkyheatmap.github.io/funkyheatmapjs/index.html) can be used to create visualizations that can be interactively browsed and sorted by a selected metric. 
 However, it was not designed to accommodate in-depth exploration via interactive aggregation or re-weighting of metrics by the user, highlighting a fundamental difference in design philosophy compared to _bettr_. 
 Benchmarking platforms such as [OpenEBench](https://openebench.bsc.es/) [@capella2017-openebench] and [OpenProblems](https://openproblems.bio/) [@Luecken2025-openproblems] also produce result visualizations for the included benchmarks, with OpenProblems using funkyheatmap for this purpose. 
 These again serve a different use case than _bettr_: by being fully integrated into the respective platforms, they provide consistent visualizations for all included benchmarks, but are not as easily accessible to a user aiming to visualize the results of their own study. 
